refactor(components): Delegate viewer functionalitty into mixin
