+++ packages/components/src/FigspecViewer/utils.ts
@@ -1,4 +1,7 @@
 import * as Figma from "figma-js";
+import { CSSResultArray, LitElement } from "lit-element";
+
+export type SizedNode = Extract<Figma.Node, { absoluteBoundingBox: any }>;
 
 export interface Point2D {
   x: number;
@@ -167,3 +170,22 @@ export function round(n: number) {
  * }
  */
 export type Constructor<T> = new (...args: any[]) => T;
+
+export function extendStyles(
+  left: typeof LitElement.styles,
+  right: typeof LitElement.styles
+): CSSResultArray {
+  return [...stylesToArray(left), ...stylesToArray(right)];
+}
+
+function stylesToArray(styles: typeof LitElement.styles): CSSResultArray {
+  if (!styles) {
+    return [];
+  }
+
+  if (styles instanceof Array) {
+    return styles;
+  }
+
+  return [styles];
+}
